pdf_read_final/requirements.txt
fastapi
uvicorn
pdfplumber
pytesseract
opencv-python
Pillow
python-dotenv
numpy